secgen/requirements.txt
pysarif
python-dotenv
click
openai
ollama
requests
pysarif